feat: insert deleteFile funcion at the DeleteProductService

+++ backend/src/services/product/DeleteProductService.ts
@@ -1,13 +1,27 @@
-import { ProductRepository } from "../../repositories";
+import { ProductRepository, OrderProductsRepository } from "../../repositories";
+import { deleteFile } from "../../utils";
 
 class DeleteProductService {
      async execute(id: string) {
 
           const repo = ProductRepository;
+          const orderProductRepo = OrderProductsRepository;
+
+          const productExistsAtSomeOrder = await orderProductRepo.findUnique({
+               where: {id}
+          });
+
+          if(productExistsAtSomeOrder) return new Error("Um produto listado em um pedido não pode ser excluído.");
 
           const product = await repo.findUnique({
                where: {id}
           });
+
+          if(product?.productImage) {
+               const fileName = product.productImage.split("/").pop();
+               const filePath = `produtos/${id}/${fileName}`;
+               await deleteFile(filePath);
+          }
 
           await repo.delete({
                where: {id}
